+++ configuration/environment_development.php
@@ -18,6 +18,7 @@ return [
     ],
 
     'application' => [
+        'cache_path' => $dotenv->pop('APP_CACHE_PATH', $app_path . '/cache'),
         'demo' => filter_var($dotenv->pop('APP_DEMO'), FILTER_VALIDATE_BOOLEAN),
     ],