+++ configuration/environment_production.php
@@ -19,6 +19,7 @@ return [
     ],
 
     'application' => [
+        'cache_path' => $dotenv->pop('APP_CACHE_PATH', $app_path . '/cache'),
         'demo' => filter_var($dotenv->pop('APP_DEMO'), FILTER_VALIDATE_BOOLEAN),
     ],