+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog of flusio
 
+## xxxx-yy-zz - v0.2 (dev)
+
+### Migration notes
+
+(optional) The responses from links fetching are now cached during one day. The
+`APP_CACHE_PATH` env variable can be set in order to change the location of
+cached responses (default is the flusio `cache/` folder).
+
 ## 2020-06-25 - v0.1
 
 First version